# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para el asistente conversacional de trámites públicos en Bolivia.",
        "title": "Mi Trámite Bolivia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/tramites": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Inserta un nuevo trámite en la BD y genera su embedding vectorizado. Requiere JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tramites"
                ],
                "summary": "Crear un trámite",
                "parameters": [
                    {
                        "description": "Datos del trámite",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTramiteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/chat": {
            "post": {
                "description": "Busca el contexto en PostgreSQL (pgvector) y genera una respuesta con Gemini.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Chat interactivo RAG",
                "parameters": [
                    {
                        "description": "Consulta del usuario",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Autentica a un administrador y retorna un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login de administrador",
                "parameters": [
                    {
                        "description": "Credenciales de administrador",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ChatRequest": {
            "type": "object",
            "required": [
                "dispositivo_id",
                "pregunta"
            ],
            "properties": {
                "dispositivo_id": {
                    "type": "string"
                },
                "pregunta": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateTramiteRequest": {
            "type": "object",
            "required": [
                "documento_contexto",
                "nombre"
            ],
            "properties": {
                "costo": {
                    "type": "string"
                },
                "documento_contexto": {
                    "type": "string"
                },
                "horario": {
                    "type": "string"
                },
                "institucion_id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "palabras_clave": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "requisitos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "nombre_usuario",
                "password"
            ],
            "properties": {
                "nombre_usuario": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}